Fix case_INTEGRATION trace for A/B two-phase ingress
- workload.c: one MC dispatch is now A-tile + B-tile = 2*CMD_PKT packets.
  Advance N every 2*CMD_PKT; latch both the A-phase first packet
  (packet_in_tile==0) and B-phase first packet (==CMD_PKT); label A/B
  from the header is_b (pe_cfg_length[15]); mask length[4:0]; double the
  expected-packet count; flush prints cur_m-1 (cur_m already advanced).
- tb.h / tb.cpp: intg_get_pe_cfg_length now returns uint16_t. It was
  uint8_t, which truncated the 16-bit field and dropped the is_b tag at
  bit15 (so B packets were mislabeled as IFMAP A).

Trace now labels IFMAP A / Filter B correctly; 256/256 packets, M=0/1.

+++ tb/testbench/dla/tb.cpp
@@ -265,7 +265,7 @@ uint8_t  ctrl_get_global_flush(void) { return top->obs_global_flush; }
 void intg_set_mock_pe_cfg_ready(uint8_t val) { top->mock_pe_cfg_ready = val; }
 void intg_set_mock_pe_data_ready(uint8_t val) { top->mock_pe_data_ready = val; }
 uint8_t  intg_get_pe_cfg_valid(void) { return top->obs_pe_cfg_valid; }
-uint8_t  intg_get_pe_cfg_length(void) { return top->obs_pe_cfg_length; }
+uint16_t intg_get_pe_cfg_length(void) { return top->obs_pe_cfg_length; } // [Iris] uint8_t→uint16_t:pe_cfg_length 是 16-bit,bit15 帶 A/B tag,uint8_t 會截掉
 uint32_t intg_get_pe_cfg_bitmask(void) { return top->obs_pe_cfg_bitmask; }
 uint8_t  intg_get_pe_data_valid(void) { return top->obs_pe_data_valid; }
 uint32_t intg_get_pe_data_nzvalue(void) { return top->obs_pe_data_nzvalue; }
